+++ src/resources/views/products/view.blade.php
@@ -12,8 +12,8 @@
         <div class="col-md-12">
             <div class="nav-controls text-right">
                 <div class="btn-group" role="group">
-                @if (count($products) > 0)
-                <a href="" class="btn btn-default btn-sm disabled btn-text">{{ $products->firstItem() . ' to ' . $products->lastItem() . ' of ' . $products->total() }}</a>
+                @if (count($models) > 0)
+                <a href="" class="btn btn-default btn-sm disabled btn-text">{{ $models->firstItem() . ' to ' . $models->lastItem() . ' of ' . $models->total() }}</a>
                 @endif
                 {!! HTML::link('admin/products/create', Lang::get('redminportal::buttons.create_new'), array('class' => 'btn btn-primary btn-sm')) !!}
             </div>
@@ -21,24 +21,25 @@
         </div>
     </div>
 
-    @if (count($products) > 0)
+    @if (count($models) > 0)
         <table class='table table-striped table-bordered table-condensed'>
             <thead>
                 <tr>
-                    <th>{{ Lang::get('redminportal::forms.name') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.category') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.sku') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.price') }}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/products', 'name', $sortBy, $orderBy) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/products', 'category_name', $sortBy, $orderBy, trans('redminportal::forms.category')) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/products', 'sku', $sortBy, $orderBy) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/products', 'price', $sortBy, $orderBy) !!}</th>
                     <th>{{ Lang::get('redminportal::forms.summary') }}</th>
                     <th>{{ Lang::get('redminportal::forms.tags') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.featured') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.active') }}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/products', 'featured', $sortBy, $orderBy) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/products', 'active', $sortBy, $orderBy) !!}</th>
                     <th>{{ Lang::get('redminportal::forms.variation') }}</th>
+                    
                     <th></th>
                 </tr>
             </thead>
             <tbody>
-            @foreach ($products as $product)
+            @foreach ($models as $product)
                 <tr>
                     <td>{{ $product->name }}</td>
                     <td>{{ $product->category->name }}</td>
@@ -102,7 +103,7 @@
             </tbody>
         </table>
         <div class="text-center">
-        {!! $products->render() !!}
+        {!! $models->render() !!}
         </div>
     @else
         <div class="alert alert-info">{{ Lang::get('redminportal::messages.no_product_found') }}</div>